Convert scanner/parser to return symbol_type objects
In what may or may not have been a good idea, I modified the Sheep.yy bison file such that it will use "symbol_type" structs when parsing tokens instead of just straight integers.
This ostensibly will make parsing files more orderly/encapsulated, but it was a big pain to make it work!