Wrapped Barn file management within a more general Asset Manager class.
The idea here is that we can request any runtime asset object from the asset manager. The asset manager will maintain a prioritized list of paths on the file system to search for the asset. If no asset is found, it will then search any Barn files that are currently loaded for the asset. Therefore, assets on the disk take priority over assets in barn files, which is consistent with the original GK3 game.