SysFuncs are stored in vector, but keyed by a map structure to aid in quick lookup by either name or hash. SysImports are also hashed using the function signature, so we could conceivably support system functions with the same name, but different argument counts or types with a little bit more effort.
SysFuncDecl is now defined as a sub-class of SysImport, which is very important - it enables us to find SysImports saved in scripts in our system (as SysFuncDecl) and thus retrieve important metadata, like whether the function is waitable or dev-only. This allows us to change these parameters without requiring a script recompile.

Finally, defined a number of husk functions for SheepAPI calls that were emitting "not found" errors in output.