Big overhaul to scene init "SIF" logic.
Created SceneInitFile (meant as a replacement for SIF class). Main difference is that SceneInitFile is "just data" - it parses the data from the asset, but leaves it up to someone else to decide what should or shouldn't be used during the current timeblock and game state. SceneData now figures this out at runtime. I think this is closer to how the original game actually did it, and it makes sense to treat the SIF as pure data and let the rest of the game use or discard as it sees fit.