Fixed a subtle bug where using "EOF" to decide when to stop reading caused the last item from a stream to be read in twice. Particularly for Sheep "Not" instruction, this was a problem (because "Not Not" is exactly opposite of "Not").
This problem occurs and is a common misstep in C++ istream logic because you only want to read in if not EOF, but then you ALSO only want to use a read-in value if a problem (like EOF) was not encountered during the read. EOF is not set when you are at the EOF, it is set when you try to read past EOF. The result here is that you can accidentally read and use one extra value before EOF is properly set. In this scenario, it caused the last Sheep instruction to be read in twice (Sheep "Not" instruction) and totally mess up the logic.

See https://stackoverflow.com/questions/4533063/how-does-ifstreams-eof-work for some more info on this.

Because I don't use streams directly, but rather through a BinaryReader interface, I can't quite use the idiom described in the above link. However, I can simply check whether the stream is "still good" after reading a value. I simplified the interfaces of BinaryReader/Writer to just have an "OK" function - basically, you can only read/write if this returns true, and you should only use a value returned by the reader if it's still true after reading! Having two functions for CanRead/IsEof just makes things unclear.

The correct way to read until EOF is now illustrated in SheepVM. Note that you really only have to do this if you don't know where the end of the stream will be. Most of the time, the file format makes it pretty clear what the byte stream will be.