Fairly large refactor to GKActor and animation subsystems!
Renamed AnimationPlayer and VertexAnimationPlayer to Animator and VertexAnimator respectively. This is both to have less verbose names AND to match the naming used by the original game.

Reverted hierarchy change to split GKActor into GKObject/GKProp/GKActor - in hindsight, this introduced a lot of complexity. Maybe we'll still need something like this in the future, but for now, I didn't like it.

Animation/walker systems, and how they interact with GKActor, has been heavily ripped apart and rewritten. The GKActor now tracks and manages separate mesh and walker actors. This is primarily to deal with some of the complexities of the GK3 animation system that couldn't be handled by simply having a single actor with a Walker/MeshRenderer on it.