Overhauled how Shaders and Materials interact with one another. Materials now hold all values for uniforms in a shader, and MUST be activated prior to rendering with them. It's now strongly discouraged to set uniforms directly on shaders - use Materials for this purpose.
Built-in uniforms that are set for all shaders are now prefixed with "g" and include things like view/proj matrices. User-defined uniforms should be prefixed with a "u".

The Material interface is now also easier to use and less prone to temporal coupling (e.g. must call function X before Y). That being said, you must set uniforms before activating a Material for the new value to be used.