Changed lightmap textures to use bilinear filtering (we want them blurry) and clamped wrap mode (makes shadows show up correctly).
At this point, the BIG problem with lightmaps is that they don't seem to work correctly with the BSP rendering system. This begs the question...were they rendered using a separate pass? Or some weird/novel rendering technique that I need to figure out? Or were the developers NOT using BSP rendering, despite using BSP files to define the level vertices???

We'll have to figure that out next time...