A few layer improvements.
First, dedicated Pushed/Popped functions to make Enter/Exit calls clearer. Enter/Exit now also get a pointer to the layer coming from or going to, which can be used to inform layer logic.

Second, flipped layer audio saving on its head after struggling with it a bit: instead of a layer saving its ambient state on exit, layers now save ambient state on enter and restore it on exit. I thiiiink this ends up being cleaner because it's usually the layer we're entering that has strong opinions about what should or should not be saved from the previous layer.