Added GKProp class, introduced VertexAnimParams, and improved vertex animation sampling.
A few big things in this commit - first, the GKActor class was getting too large and clearly represented two distinct objects (Props vs. Actors). So, I introduced a base class called GKProp that is for props. GKActor now subclasses from this (since an Actor is, in a lot of ways, a Prop with more functionality). I'm not 100% sure I'm happy with this approach, and perhaps someday I'll revisit and consider a component-based approach or something along those lines.

Also, because the argument list for playing vertex animations was becoming quite large, I bit the bullet and added a VertexAnimParams struct. This simply contains all the various arguments you might want to pass for starting a vertex animation.

Finally, vertex animation sampling was pretty jank: you'd specify a frame to sample, but this would be converted to a time, and then back to a frame to finally do the sample. To fix this, I introduced alternate functions that take a frame as an argument instead of a time. This allows code paths that are interested in sampling a specific frame (rather than a time) to use simpler logic and avoid converting back and forth a bunch of times.