Added some custom memory allocators to the codebase.
These aren't currently used by the game in any way, and I'm not sure when exactly I'll pull the trigger on refactoring existing code to use these. But it was something that I was eager to wrap my head around, so I went ahead and read up and built a couple.