In what I _hope_ is a step in the right direction, I updated Model and VertexAnimation loading code to just use the provided transform directly, rather than swapping Y/Z values.
As I was attempting to get some lighting working with normals, I realized I was overthinking the mesh importing logic. While it is true that GK3 models appear to be Z-Up and the game is Y-Up, it turns out that the transform matrices stored with the models/animations already take that transformation into account! I was sort of manually mimicking it by flipping the Y/Z axes of the transforms, and then flipping the Y/Z coordinates of points/normals, but this was not needed.

I was also doing some totally unneeded busy work importing data into vectors/quaternions and then creating a Matrix4 from that...you can just use the Matrix4 values directly. Silly mistake when I was, again, maybe overthinking things earlier in the project.