Fixed a Quaternion->Heading bug that was messing up "Walk to Anim" approaches.
Converting a quaternion to Euler angles, and then using the y rotation, did not work correctly for calculating a heading sometimes. The reason is that a rotation represented by a quaternion can have multiple Euler angle representations. So, if we have a quaternion rotation for "180 degrees about the Y-axis", the Euler angle representation could have rotations about X/Z and no rotation about Y. To fix this, we need to calculate a heading from a quaternion using a different method.