Some GKActor animation/rendering cleanup/consolidation.
The "identifier" for a GKActor is removed - it was simply the model name, which can be retrieved with GetModelName. Model rotation variable removed from GKProp because...it was only used by subclasses - no need for it!

Animation logic in GKActor was consolidated to a couple shared functions. These changes also fix issues where GKActors would not be positioned correctly when switching scenes.