Fixed some instances where topics would appear more often than expected.
This was kind of surprising. It appears that topic case logic implicitly checks whether the topic has been discussed before, and if so, fails the case logic. Of course, there is a big exception: if the case logic EXPLICITLY checks topic count, then don't do the implict check! This requires some annoying string manipulation, but it seems to work fine for now.