Started working on logic for changing timeblocks.
I have found no data-driven way that GK3 stores this logic, but it's definitely based on flags and/or noun/verb/topic counts. So, I thought it'd be appropriate/fun to implement this timeblock change logic in a special Sheep script.

Obviously it's pretty annoying to have to figure out the exact conditions for changing a timeblock, but there are plenty of clues to this, and you can basically just play the game to figure it out!