Fairly large changes to the Sheep VM to enable specifying tags associated with running SheepScripts. The tag is mainly used as a way to identify executing SheepScripts to stop prematurely. This is now used, when popping a layer, to prematurely stop execution of any Sheep associated with that layer.
In GK3, an executing Sheep is "attached" to a particular layer. This is usually the layer that is on top of the layer stack when the Sheep starts executing. However, in some cases, it can also be set to the "global" layer.

The layer the executing Sheep is attached to controls its lifetime. For example, if an executing Sheep attached to the current scene layer is waiting for 100 seconds, but we switch scenes, that Sheep is stopped prematurely, and any code after the timer finishes is never executed. On the other hand, if a Sheep is attached to the global layer, it will finish, and changing scenes will not cause it to stop prematurely.

There's still one big change to make related to callbacks to a SheepThread if the Sheep has stopped executing, but this is a good spot to checkpoint for now.