Fixed bug where Vertex Animations would not properly update if the owning Actor was disabled. This resolves some instances where animations that are supposed to play in sync with one another would be out of sync (e.g. maid using key to open doors).
This bug was a bit tricky, since it relates to the foundational Actor/Component behavior and how Vertex Animations were implemented. When a 3D model is "hidden", its entire Actor is set to Inactive. This hides the 3D model, but also stops the entire Actor (and any Components attached) from being updated. This seems safest, as I don't want an invisible object to be accidentally clickable, or to have it update in unanticipated ways.

However, I realized that GK3 expects Vertex Animations to proceed even when a model is hidden. Since I implemented the Vertex Animator as a Component, this means the Vertex Animator does not update when the model is hidden. But when the model is unhidden, the animation continues from where it left off, and is desynchronized from other animations that were meant to start at the same time.

One way to fix this is to overhaul how Vertex Animations work. Instead of using a Component on the Actor, a higher-level manager approach could be taken. There could actually be several benefits to this long-term (e.g. being able to better orchestrate how disparate animations relate to one another). But for now, this seemed like a lot of work.

What I landed on for now is to utilize a timer/stopwatch that records how much time passes while the Vertex Animator is disabled. When the Component is re-enabled, we can "catch up" the animation by skipping forward however much time. One nice thing about this approach is that it means we aren't doing unnecessary work to animate objects that aren't visible.