Fixed a bug where 3D sound effects would sometimes be silent. Affected sounds were set to play at a specific model's position. This was resolved by updating how a model's position was calculated.
Because of how many GK3 models are authored, the world position of the model doesn't necesarrily align with the visual position of the model. For example, a model may be positioned at the world origin, but all its vertices/faces are offset significantly to position the *visuals* elsewhere in the world.

To fix this, instead of using the object's world position, we get the mesh renderer and use it's AABB center point. This seemed to improve the situation significantly.