Implemented "animation parenting," which appears to be a thing in the original game. Fixed issues with maid's spray-bottle and rag animations not playing correctly.
In short, if a model is animated as part of certain animations, we can "detect" that it should parent itself to a parent object for the duration of the animation. For example, when playing the maid's cleaning animation, the spray bottle and wiping rag should play their animations relative to the maid's current position.