MAJOR overhaul to how Actors handle positioning/rotating and walking.
This approach seems to give better results that are closer to the original game. I performed a lot of trial and error - modifying data files to see how the original game reacts - to verify that this code is behaving similarly.

It is surely not perfect yet (particularly when skipping walks, rotations can be a bit off), but it seems like a better approach to continue improving.