Fixed issue where Grace walk behavior was incorrect.
This was tricky to hunt down, and the fix is pretty obscure (and what I'd consider a hack). Most walk animation files list the DOR model animation first, then the character animation second. In Grace's case, the order was flipped. As a result, Grace would start animating before the DOR, which led to incorrect facing direction calculations. There are probably a lot of ways to fix this - changing the animation system, changing the walk/facing direction logic, etc. But I went with this approach because it's simpler in comparison.