Fixed bug where Montreaux doesn't play an animation correctly during a cutscene in the Chateau de Serras attic.
The original game's data file actually has a typo in it! The specified vertex animation doesn't exist (it has one too many letters). The original game somehow accounts for this or has the error resiliency, but it really seems to me like the right solution is to fix the data file. So, I have copied the asset here so it will override the original asset.