Fairly larger REFACTOR - Moved away from "Services" paradigm in favor of globals, enabled a lot more MSVC warnings (and fixed a lot of them too).
This change was a long time coming. Early on in development, I used the "Services" interface to store and retrieve managers in a map-based system. However, as I've gotten more experienced with C++, I've found this to just be "global access with more overhead." The "Services" approach also leads to an undesirable situation where many headers are being included transitively rather than directly. The "Services" system may still prove useful in the future if any runtime interface swapping is needed, but for now it is no longer used.

On the Visual Studio side, a great deal of warnings were enabled with the /Wall compiler flag. Then, several of them were manually disabled if they were deemed unuseful. However, the new warnings that were left enabled have helped to identify a variety of issues in the code. In at least one case, an obvious bug was fixed. The compiler now emits a lot more warnings, but these will be fixed in future commits.