Changed LocationManager to not change location right away when a location change is requested, but instead wait until the next Update call.
This ensures location changes always occur at a known/specific time during the update loop. It also allows any in-progress Action to continue/finish before processing the location change request. This is important if logic for changing a timeblock depends on a variable that isn't incremented until the current action has finished.