Fixed bug where kitchen music would continue to play after leaving the kitchen.
After comparing to the original game, it seems that the Soundtrack "StopMethod" default value of "PlayToEnd" is NOT respected when a scene change occurs; it's only respected if you stop a soundtrack via a stop command in the current scene. So, I had to add some extra logic to force a "PlayToEnd" audio file to stop in that case.