Added logic to BSP raycaster to ignore hit test polygons when the ray hits a transparent pixel.
When BSP specifies hit tests, there's a chance that the hit test texture is not fully opaque and has transparent pixels. In that case, if the raycast hits a transparent pixel, it SHOULD NOT be considered a hit. These changes add that logic. In practice, this was a problem in the Church when tracing the tilted square. The four angels use hit tests that overlap one another, and the way each angel's hit test works is via a small set of opaque pixels in a larger square image. Without this change, it's impossible to click on 3 of the 4 angels.