Fixed a bug where unlocking the mouse, and then checking for mouse button trailing edge after would count as a mouse click interaction.
The problem is that "mouse lock" is used when you hold down the left mouse button to move the camera around. Upon releasing the mouse button (to stop moving the camera, and unlocking the mouse), the mouse button release also counts as a "left mouse button click" in other code. The practical negative effect is that the game mistakes you releasing the left mouse button as a scene interaction.

To fix this, we should register the intent to lock/unlock the mouse, but only actually change the lock state at a specific time at the beginning of a frame.