Added an Actor LateUpdate to better support vertex animation parenting.
I found that child objects in vertex animations would sometimes be positioned incorrectly. This happens because the parent object might not necessarily update before the child object does - it depends on the order of the Actors list. To resolve this, adding a "LateUpdate" callback allows components to run code after all other Actors have updated. This may be slightly overkill for this specific problem, but it can be useful in other situations as well, in the future.