Further refined SoundtrackPlayer stop logic to improve the behavior.
Semi-recently, I made changes to SoundtrackPlayer to ensure a scene music track ended when you left the scene (the Kitchen music) - this seemed to be a developer oversight. However, this change also caused problems elsewhere, where certain sound effects played as you left a scene were not playing correctly anymore.

To fix this, I added logic to explicitly stop scene audio when unloading the scene, including the force flag so that any "PlayToEnd" audio is also forced to stop. This only affects soundtracks defined in the SIF, and doesn't affect soundtracks played through Sheepscript. This seems to give me the needed behavior.