Further refined the logic for choosing which Action to use when there are multiple valid options.
The logic to choose which action to priotize amongs many valid ones is very tricky and complex. I discovered more instances where it was failing to choose the correct action. To fix these, I made a few changes to the logic again. For one, custom case logic is checked first always now. For two, if multiple custom cases are valid, the one that is alphabetically LATER is the one now chosen - as crazy as that sounds, I experimented with the original game, and that DOES appear to be how it works!

We'll see if this works...or if more problems are discovered and further tweaks will be needed.