Optimized walker pathfinding code by ammortizing the cost of an expensive operation.
After profiling walker code, I found that the operation of adjusting each path node's y-position during path calculation was expensive and eating a lot of time. This does involve a raycast against the scene BSP for each node, so that kind of makes sense. This is also even less ideal because a lot of those nodes are discarded shortly after doing this work!

To improve the situation, I removed the pre-calculation of path node height entirely. Path node heights are calculated only when the walker is about to walk to that path node. This ammortizes the cost of doing these raycasts over many frames, and it also avoids doing the raycast on nodes that will be pruned from the path shortly anyways.