Refactored GasPlayer to execute GAS nodes as soon as it starts playing, and to play zero-length nodes in sequence.
The original purpose of this commit was to fix an issue where some actors wouldn't have their GAS executed on the first frame of a scene, resulting in one rendered frame where they were in the wrong pose or location. This is caused by order-of-operation issues - we "Start" the Gas Player, but the first node doesn't execute until the next time OnUpdate is called. If that doesn't happen until next frame, then this frame will render incorrectly.

To fix this, when starting a GAS in a GasPlayer, at least one node is executed immediately. As a further fix, I also detected when a GAS node has zero duration and continued to play nodes until we hit one that takes up time. This fixes issues where immediate, zero-length nodes like changing or checking variables in GAS scripts would actually take at lease 0.016s, since we'd only do one per frame.

Finally, moving the GKProp StartFidget call to Init resolves some issues where a GAS node plays an animation on the first frame, but the object hasn't yet been added to the scene. We must fully construct the scene first, and then do this sort of thing in Init, where all scene objects are present to manipulate.