Went down a bit of an optimization rabbit hole with pathfinding. Introduced some optimizations and ways to reduce node counts.
For one, I did some more experimenting with A*. I found that BFS still outperformed it, but left the code here just in case. One big optimization was to only use every nth node in the grid when doing BFS - in a lot of cases, this can reduce the number of nodes needed to be explored by 75%. If a path isn't found, we do retry with a higher "fidelity," so the less optimized approach can be used worst-case.

I also added a "string pulling" algorithm that gets rid of path nodes that aren't needed. If we can walk a straight line between the node before and the node after, then the current node is unneeded.