Updated action bar positioning logic to more closely match the original game.
Up to this point, I had some logic to always position the action bar under your mouse cursor when it showed. However, in the original game, this only happens when a user action directly causes the action bar to show - if it shows automatically after an action/cutscene, it should appear in the same spot it was last located at. This commit updates our action bar logic to do just that.