Fixed walk action during action skip doing unnecessary pathfinding/animation work.
If a walk command is initiated during an action skip, it is now immediately forced to complete with no pathfinding or animation playing at all - we just put the walker at the correct position/heading. Besides being more efficient, this also fixes at least one bug where a walk during an action skip caused Gabe to be in the wrong position after the action skip was done.