Patched (pending further verification) a Mac bug where diffuse/lightmap textures would get assigned uncommon texture units, resulting in graphical artifacts throughout the game.
Blob shadows are implemented using the lightmap shader (since they are kind of like a dynamic lightmap). So they have two textures: a diffuse/color and the shadow/lightmap texture.

Because material uses an unordered_map, these two textures can theoretically be assigned to texture units in any order - as the name implies, the collection is unordered. But it just so happens on Windows that the diffuse texture was always assigned texture unit 0 and the lightmap texture was always assigned texture unit 1. On Mac, this didn't hold, and the two were swapped.

It shouldn't really be a problem for this to happen. When we assign a texture to a texture unit, we also tell the shader which texture unit is being used by that sampler. However, it clearly DOES cause graphical issues when they are swapped on Mac. The reason for this is unknown (that's a bug to fix), but it may have something to do with the BSP code for rendering lightmaps not playing nicely with the material code.

Anyway, an easy hack/band-aid here is to use an std::map so the textures are sorted consistently across all platforms. It just so happens they will sort in the correct order due to uDiffuse being before uLightmap alphabetically. Not my favorite thing, but it fixes the problem in the short-term.