Revised UI widget factory functions to take in a name for the widget. Took a pass across most UIs to refactor them to use the factory functions and to name all UI elements.
This commit accomplishes several goals. First, the API for UI helper functions is a bit cleaner and easier to use. It also requires specifying a name, which helps keep the Scene Hierarchy tool readable (it really hurts to debug and most items in the Scene Hierarchy have no name). Another thing this accomplished was forcing me to go back through old UIs and "modernize" them to use the new factory methods, which cuts down on boilerplate code in many places.