Fixed Math::Acos sometimes returning NaN, even though it theoretically should not.
When doing vector math, I encountered a problem where doing "Acos(Dot(unitVec1, unitVec2))" would return NaN. This shouldn't happen since the dot product of two unit vectors is always in the range -1 to 1, which matches the range of Acos. However, due to floating point imprecision, dotting two unit vectors could sometimes produce a vector that was *just barely* outside the value range (like 1.00000012 or similar). Acos rejects this with NaN.

To fix this, the only solution I really see is to clamp the ratio passed in to Acos (and Asin, since it has similar restrictions, though less frequently used). It's true this may clamp a completely errant passed in ratio, but it also resolves the "bug" where perfectly valid math is being performed, but you still end up with NaN unexpectedly.