Added support for a border width in UINineSlice.
When using just colors for UINineSlice center and sides, the border size defaults to 1 pixel (when you use a texture, the size is derived from the texture width/height). However, there are times when you want the border to be a specific width because it looks better. This new parameter allows that.