Fixed UI widgets not refreshing when a window resolution change occurred.
This was most problematic for the status overlay label and the debug console. In both cases, changing resolution would result in the labels being incorrectly positioned until you changed the label text. Now that labels are dirtied on screen resolution change, they very elegantly update to the correct position on screen resolution change.