Fixed toggling fullscreen and changing resolutions sometimes resulting in an unplayable game.
It seems that it's necessary to call SDL_SetWindowSize even when changing resolution in fullscreen mode. Without this, you can get some weird situations where the game is at a lower resolution than the window in fullscreen mode (or vice-versa), resulting in a weirdly cropped 3D view that's virtually unplayable.