Fixed a bug in two Sidney font files that would cause the incorrect glyph to be rendered for certain characters.
In two font files for Sidney, the list of characters in the font doesn't match associated texture. Specifically, the font asset lists four "I" glyphs with accents, but the font texture only has three. The result of this bug was that any glyph you attempted to render after the mismatch would render a different glyph (an "off by one" bug).