Fixed a bug where scene animations would still update when the scene is paused, sometimes causing 3D objects in the scene to move/animate even though the scene was paused.
A long time ago, when I implemented scene pausing when some other screen was up, I didn't pause the scene animator because it was still needed on other screens - for example, performing actions in inventory would run animations. However, more recently, I did add a "global" animator to deal with subtitle animations during videos - this one never pauses.

A problem I started to notice is that some 3D objects would suddenly move behind the inventory screen. The cause of this ultimately was that the scene animator wasn't paused, so it'd still send "anim finished" callbacks to the GAS system. This in turn might try to play a new scene animation, triggering the 3D model to be moved by a sampled animation.

To fix this, I just needed to undo my old hack - when the scene pauses, the scene animator now also pauses. One wrinkle here is that any animation that might be playing when the scene is paused needs to be sure to choose the right animator so it doesn't get stuck. This is only (I hope) dialogue YAKs, so they now use the global animator if the scene animator is paused.