Implemented missing InspectModelUsingAngle Sheep API function. Fixed edge case where INSPECT and INSPECT_UNDO actions could be shown at the same time.
This Sheep API call is only used a few times in the first timeblock of the game: when inspecting Lady Howard and Estelle, and when inspecting Emilio in the phone booth. It now works! As a bonus, this also helped uncover a bug with INSPECT/INSPECT_UNDO that was fixed.