Implemented missing autoscript nodes related to doing cleanup/interrupt anims, changing idles, and walking to positions when an autoscript is interrupted by a player action.
Though autoscripts have been working pretty well, some functionality was missing that causes certain animations to play when an autoscript is interrupted, or that causes an actor to walk to a specific spot when interrupted, or changes the idle fedget on interrupt, or even clears flags on interrupt. In turn, this led to several scenarios throughout the game with incorrect behavior. For example, Gabe when talking to Mosely in R33 wouldn't properly stand up when performing actions or sit down when talking to Mosely. When Grace talks to the Abbe in the office, her sit/walk behavior was very odd. These changes fix those issues.