Fixed a bug where various semi-transparent textures (usually foliage) would have a single opaque pixel visible in the upper-left corner of the image.
This issue was most noticeable with trees/foliage around the game. Each foliage texture has mostly transparent pixels around the edges, but pixel (0, 1) is often opaque. My guess is that this is some sort of key value that the original game might read/use for some purpose. But I haven't found a need for it, and we don't want to render it since it looks wrong.