Modified pathfinding code to return a "best effort" path in cases where pathfinding doesn't find a full path to a goal position. This fixes some "walk to see", "walk near model", and player click-to-walk bugs.
Before this change, if the pathfinder failed to find a path, it would fall back on "just walk directly from current position to goal." This would fail to be the correct behavior in several scenarios:
- If the player clicks on on a spot on the floor that isn't walkable, the character would walk directly there in a straight line, clipping through any objects in the way. Additionally, if the player clicked on a walkable spot that was unreachable, the character would walk in a straight line and clip through things.
- For "walk to see" approaches, if the walk position calculated wasn't reachable in the walk graph, the character would walk in a straight line and clip through things.
- Similar for "walk near model" if the walk position wasn't reachable on the walk graph.

To solve this, I updated the pathfinding code to always return a "best effort" path, even if the final position couldn't be reached. In this scenario, the game paths to the position we found during pathfinding that was closest to the goal. The result seems pretty smart, and causes the character to pathfind intelligently and then stop when they reach the last walkable spot on the way to the goal.