Fixed bug where Grace is not positioned correctly in R25 at the beginning of Day 3, 12PM timeblock.
The positioning code for the beginning of Day 3, 12PM relies on Grace's last location being R25 and her current location being R25. 99% of the time, the game doesn't allow that to happen. If you "SetLocation" when already at a location, last location is not updated to be the same as the current location. However, this does seem to happen and be allowed during timeblock changes.

This change fixes the bug, but will need to test all timeblock transitions to make sure it doesn't break anything.