Added support for explicitly setting a parent for animations (rather than implicitly detecting based on file name).
Fixes bug where Buchelli is supposed to be carrying a newspaper while walking, but the newspaper is missing.

This is *probably* a better approach than the implicit option, and it may be able to replace the implicit approach entirely. As far as I know, for all animations with parenting, they are going to be relative to some character (like a walk animation).