Fixed Day 1, 2PM timeblock ending prematurely if the last thing you do is follow Buthane.
A problem here was that one of the timeblock completion conditions is to follow Buthane. But if it's the last thing you do, all conditions are met, and you move to the next timeblock before even playing the follow cutscene on the map. To fix this, I implemented a workaround where following Wilkes/Buthane to their destinations sets their locations to the spot they stop at - this happens at the END of the follow cutscene. Then, we can check their location instead of their follow counts as part of the timeblock end conditions.