Fixed (I think) a pretty serious bug where custom Queue/Stack containers with more complex objects (like std::string) would corrupt memory.
I noticed that the in-game console command history was becoming corrupted and showing garbage sometimes. Upon closer inspection, the strings being inserted into the history Queue were being corrupted. After a bit of research, I realized that the method of copying an object into the container's underlying byte storage wasn't safe/correct for anything other than PODs. In this case, since std::string is not a POD, it caused problems.

The fix is actually kind of nice - we just need to use the same behavior as Emplace (which is placement new). But with the "Push" variant, we just rely on the copy constructor to be executed by "Emplace" when duping the object into the memory buffer.