Fixed oversight where autoscript "distance condition" nodes weren't checked immediately upon encountering them, resulting in the potential to errantly play wrong autoscript nodes.
This fixes a bug where Madeline's animations at CSD during Day 1, 2PM timeblock were incorrect. This ended up being a pretty subtle issue - in one case, an autoscript adds a distance condition based on whether Gabriel is nearby. If Gabriel is already nearby, the old code would wait one frame before doing the distance check - in the meantime, the autoscript player would start an animation it shouldn't have. When the distance was then checked, it would play a different anim, interrupting the first one in a bad way.

Another way to look at this is, if a distance condition immediately evaluates to true when it's encountered in an autoscript, we need to take the branch immediately instead of allowing nodes that wouldn't play if the condition was true to play.