Updated how the "talkto" action attribute is used. It will now actually execute any TALK action that exists for a character, which seems to fix some conversation issues.
Certain actions in GK3 have a "talkto" attribute that isn't particularly well documented. Via observation, it seems to mean that after the action finishes, the topic bar should open so you can talk to someone.

However, in at least one case (Grace/Simone talking in Day 2, 5PM), the resulting behavior was incorrect. Somehow, Grace/Simone were having a conversation set on them with SetConversation, which wasn't happening in the "walk up to desk" code. The only time the conversation was set was in the "SIMONE/TALK" action.

So, this got me thinking that perhaps the "talkto" attribute is meant to execute the "TALK" verb on a the noun stored in "talkto" after the current action completes. That's what this code change does, and it does seem to fix the Simone/Grace issue. I hope it doesn't break anything elsewhere...