Fixed oversight where Sidney Analyze Map layout did not match the original game.
Much to my horror, I realized that the map layout I had created was slightly off from the original game. In particular, the square around the circle was at the wrong angle. This in turn led most other points/shapes placed afterwards to be wrong too. While it wasn't hugely wrong, it was enough to be noticeable especially when placing "The Site" on the map.

Anyway, in this commit I bit the bullet and fixed the incorrect layout. I also added some code that should fix old save games to have the correct layout when they are loaded.